requirements.txt
Appium-Python-Client
pillow
pyssim
